refactor(core): extract shared token-mode discovery manifest (#546)
